refactor(web): share lazy connector definitions
