+++ apps/web/src/lib/wallet/namespaces/evm/adapters/safe.ts
@@ -2,19 +2,19 @@
 
 import type { CreateConnectorFn } from '@wagmi/core'
 import { getConnectorById } from '../utils/connector'
+import { safeConnectorDefinition } from './safe-definition'
 
 let createConnectorFn: CreateConnectorFn | undefined
 
 async function getCreateConnectorFn() {
   if (createConnectorFn) return createConnectorFn
 
-  const { safe } = await import('@wagmi/connectors')
-  createConnectorFn = safe()
+  createConnectorFn = await safeConnectorDefinition.load()
   return createConnectorFn
 }
 
 function getConnector() {
-  return getConnectorById('safe')
+  return getConnectorById(safeConnectorDefinition.id)
 }
 
 export const getSafeConnector = async () => {
