+++ apps/web/src/lib/wallet/namespaces/evm/adapters/walletconnect.ts
@@ -1,23 +1,20 @@
 'use client'
 
 import type { CreateConnectorFn } from '@wagmi/core'
-import { WALLET_CONNECT_PROJECT_ID } from 'src/lib/wagmi/config/connector-options'
 import { getConnectorById } from '../utils/connector'
+import { walletConnectConnectorDefinition } from './walletconnect-definition'
 
 let createConnectorFn: CreateConnectorFn | undefined
 
 async function getCreateConnectorFn() {
   if (createConnectorFn) return createConnectorFn
 
-  const { walletConnect } = await import('@wagmi/connectors')
-  createConnectorFn = walletConnect({
-    projectId: WALLET_CONNECT_PROJECT_ID,
-  })
+  createConnectorFn = await walletConnectConnectorDefinition.load()
   return createConnectorFn
 }
 
 function getConnector() {
-  return getConnectorById('walletConnect')
+  return getConnectorById(walletConnectConnectorDefinition.id)
 }
 
 export const getWalletConnectConnector = async () => {
